Implemented a base class for ACSD and mutual information
